add test for TestGetValueByNamespace()
Prior to this commit, no tests existed for TestGetValueByNamespace(). I
discovered some behavior where, it appears, that TestGetValueByNamespace
is returning the memory address of a variable as a string, instead of
the value itself. For example:

  * /Users/roger/work/gocode/src/github.com/intelsdi-x/snap-plugin-utilities/ns/namespace_test.go
  Line 1318:
  Expected: '3.14'
  Actual:   '0xc820181948'
  (Should be equal)

This commit adds a test for GetValueByNamespace() that reproduces this
issue in order to patch GetValueByNamespace() to correct this behavior.