handle pointers in GetValueByNamespace()
Prior to this commit, GetValueByNamespace() couldn't handle pointers
(see commit message in bca93f4). Specifically, this would happen:

  * /Users/roger/work/gocode/src/github.com/intelsdi-x/snap-plugin-utilities/ns/namespace_test.go
  Line 1318:
  Expected: '3.14'
  Actual:   '0xc820181948'
  (Should be equal)

This commit adds an additional case before the default case for handling
type reflect.Ptr. If the type is a pointer, val is set to the value that
the pointer points to. It then returns if len(ns) is 1, or goes deeper
if the length of the namespace is greater than 1.